feat: [CDS-73657]: Add webhook event directly into queue without processing from SCM service
