+++ 953-events-api/src/main/java/io/harness/eventsframework/EventsFrameworkConstants.java
@@ -34,6 +34,7 @@ public final class EventsFrameworkConstants {
   public static final String GIT_PUSH_EVENT_STREAM = "git_push_event_stream";
   public static final String GIT_PR_EVENT_STREAM = "git_pr_event_stream";
   public static final String WEBHOOK_PUSH_EVENT = "WebhookPushEvent";
+  public static final String WEBHOOK_TRIGGER_EVENT = "WebhookTriggerEvent";
   public static final String WEBHOOK_BRANCH_HOOK_EVENT = "WebhookBranchHookEvent";
   public static final String WEBHOOK_EVENT = "WebhookEvent";
   public static final String GIT_BRANCH_HOOK_EVENT_STREAM = "git_branch_hook_event_stream";
